enh(Data): move SQLExecutor to a library in Poco::Data testsuite #4230
