+++ Data/ODBC/Makefile
@@ -12,8 +12,9 @@ objects = Binder ConnectionHandle Connector EnvironmentHandle \
 	Extractor ODBCException ODBCMetaColumn ODBCStatementImpl \
 	Parameter Preparator SessionImpl TypeInfo Unicode Utility
 
+target_includes = $(POCO_BASE)/Data/testsuite/include
 ifndef POCO_DATA_NO_SQL_PARSER
-	target_includes = $(POCO_BASE)/Data/src
+	target_includes += $(POCO_BASE)/Data/src
 endif
 
 target         = PocoDataODBC
