+++ Data/ODBC/testsuite/Makefile
@@ -33,12 +33,13 @@ ifeq ($(POCO_CONFIG),MinGW)
 	objects += ODBCAccessTest
 endif
 
+target_includes = $(POCO_BASE)/Data/testsuite/include
 ifndef POCO_DATA_NO_SQL_PARSER
-	target_includes = $(POCO_BASE)/Data/src
+	target_includes += $(POCO_BASE)/Data/src
 endif
 
 target         = testrunner
 target_version = 1
-target_libs    = PocoDataODBC PocoData PocoFoundation CppUnit
+target_libs    = PocoDataODBC PocoDataTest PocoData PocoFoundation CppUnit
 
 include $(POCO_BASE)/build/rules/exec
