+++ Data/testsuite/Makefile
@@ -1,30 +1,11 @@
 #
 # Makefile
 #
-# Makefile for Poco Data testsuite
+# Makefile for Poco Data Tests
 #
 
-include $(POCO_BASE)/build/rules/global
-
-ifeq ($(findstring SunOS,$(POCO_HOST_OSNAME)),SunOS)
-	POCO_SUN_FORTE = $(findstring SunOS-SunForte, $(POCO_CONFIG))
-	POCO_SUN_STUDIO = $(findstring SunOS-SunStudio, $(POCO_CONFIG))
-	ifneq (,$or ($(POCO_SUN_FORTE), $(POCO_SUN_STUDIO)))
-		CXXFLAGS += -erroff=hidevf
-	endif
-endif
-
-objects = DataTestSuite Driver \
-	DataTest SessionPoolTest \
-	Binder Extractor Preparator SessionImpl Connector TestStatementImpl
-
-ifndef POCO_DATA_NO_SQL_PARSER
-	objects += SQLParserTest
-	target_includes = $(POCO_BASE)/Data/src
-endif
-
-target         = testrunner
-target_version = 1
-target_libs    = PocoData PocoFoundation CppUnit
-
-include $(POCO_BASE)/build/rules/exec
+.PHONY: projects
+clean distclean all: projects
+projects:
+	$(MAKE) -f Makefile-testrunner $(MAKECMDGOALS)
+	$(MAKE) -f Makefile-library $(MAKECMDGOALS)
