+++ Data/testsuite/Makefile-library
@@ -1,0 +1,21 @@
+#
+# Makefile
+#
+# Makefile for Poco Data testsuite library
+#
+
+include $(POCO_BASE)/build/rules/global
+
+objects = SQLExecutor
+
+target_includes = $(POCO_BASE)/Data/testsuite/include
+ifndef POCO_DATA_NO_SQL_PARSER
+	objects += SQLParserTest
+	target_includes += $(POCO_BASE)/Data/src
+endif
+
+target         = PocoDataTest
+target_version = 1
+target_libs    = PocoData PocoFoundation CppUnit
+
+include $(POCO_BASE)/build/rules/lib
